Dividing logic into handlers
